backend/requirements.txt
# Agent framework
openclaw>=0.5.0


# API server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0

# Data models
pydantic>=2.6.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Paper trading
alpaca-py>=0.20.0

# Market data (optional enrichment)
yfinance>=0.2.37
pandas>=2.2.0

# Logging & storage
aiosqlite>=0.20.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
